Put services in namespaces, renamed webots/gazebo bridge
