+++ eager_bridge_webots/src/eager_bridge_webots/webots_bridge.py
@@ -3,7 +3,7 @@ import functools
 import re
 from eager_core.physics_bridge import PhysicsBridge
 from eager_core.utils.file_utils import substitute_xml_args
-from eager_core.utils.message_utils import get_value_from_def, get_message_from_def, get_response_from_def
+from eager_core.utils.message_utils import get_value_from_def, get_message_from_def, get_response_from_def, get_length_from_def, get_dtype_from_def
 from webots_ros.msg import Float64Stamped
 from webots_ros.srv import set_int, set_float
 from sensor_msgs.msg import Image
@@ -72,7 +72,7 @@ class WeBotsBridge(PhysicsBridge):
             sensor = sensors[sensor_name]
             topic_list = sensor['names']
             space = sensor['space']
-            robot_sensors[sensor_name] = [get_value_from_def(space)]*len(topic_list)
+            robot_sensors[sensor_name] = [get_value_from_def(space)] * get_length_from_def(space)
             for idx, webots_sensor_name in enumerate(topic_list):
                 enable_sensor = rospy.ServiceProxy(name + "/" + webots_sensor_name + "/enable", set_int)
 
@@ -90,11 +90,12 @@ class WeBotsBridge(PhysicsBridge):
                     else:
                         self._sensor_subscribers.append(rospy.Subscriber(name + "/" + webots_sensor_name + "/value",
                             Float64Stamped, functools.partial(self._sensor_callback, name=name, sensor=sensor_name, pos=idx)))
-            self._sensor_services.append(rospy.Service(topic + "/" + sensor_name, get_message_from_def(space), functools.partial(self._service,
-                                                                                                buffer=self._sensor_buffer,
-                                                                                                name=name,
-                                                                                                obs_name=sensor_name,
-                                                                                                message_type=get_response_from_def(space))))
+            self._sensor_services.append(rospy.Service(topic + "/sensors/" + sensor_name, get_message_from_def(space),
+                                                       functools.partial(self._service,
+                                                                         buffer=self._sensor_buffer,
+                                                                         name=name,
+                                                                         obs_name=sensor_name,
+                                                                         message_type=get_response_from_def(space))))
         self._sensor_buffer[name] = robot_sensors
     
     def _init_actuators(self, topic, name, actuators):
@@ -107,7 +108,7 @@ class WeBotsBridge(PhysicsBridge):
             for webots_actuator_name in topic_list:
                 set_action_srvs.append(rospy.ServiceProxy(name + "/" + webots_actuator_name + "/set_position", set_float))
 
-            get_action_srv = rospy.ServiceProxy(topic + "/" + actuator_name, get_message_from_def(space))
+            get_action_srv = rospy.ServiceProxy(topic + "/actuators/" + actuator_name, get_message_from_def(space))
             robot_actuators[actuator_name] = (get_action_srv, set_action_srvs)
         self._actuator_services[name] = robot_actuators
 
@@ -117,18 +118,19 @@ class WeBotsBridge(PhysicsBridge):
             state = states[state_name]
             topic_list = state['names']
             space = state['space']
-            robot_states[state_name] = [get_value_from_def(space)]*len(topic_list)
+            robot_states[state_name] = [get_value_from_def(space)] * get_length_from_def(space)
             # for idx, webots_sensor_name in enumerate(topic_list):
             #     enable_sensor = rospy.ServiceProxy(name + "/" + webots_sensor_name + "/enable", set_int)
             #     success = enable_sensor(self._step_time)
             #     if success:
             #         self._sensor_subscribers.append(rospy.Subscriber(name + "/" + webots_sensor_name + "/value",
             #             Float64Stamped, functools.partial(self._sensor_callback, name=name, sensor=sensor, pos=idx)))
-            self._sensor_services.append(rospy.Service(topic + "/" + state_name, get_message_from_def(space), functools.partial(self._service,
-                                                                                                buffer=self._state_buffer,
-                                                                                                name=name,
-                                                                                                obs_name=state_name,
-                                                                                                message_type=get_response_from_def(space))))
+            self._sensor_services.append(rospy.Service(topic + "/states/" + state_name, get_message_from_def(space),
+                                                       functools.partial(self._service,
+                                                                         buffer=self._state_buffer,
+                                                                         name=name,
+                                                                         obs_name=state_name,
+                                                                         message_type=get_response_from_def(space))))
         self._state_buffer[name] = robot_states
 
     def _sensor_callback(self, data, name, sensor, pos):
