+++ eager_bridge_webots/src/eager_bridge_webots/webots_node.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import rospy
-from eager_bridge_webots.webots import WeBotsBridge
+from eager_bridge_webots.webots_bridge import WeBotsBridge
 
 if __name__ == '__main__':
